requirements.txt
altair==5.5.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.46
botocore==1.38.46
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
entrypoints==0.4
et_xmlfile==2.0.0
Faker==37.4.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.58.4
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
htbuilder==0.9.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==6.0.0
Markdown==3.8.2
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.44.0
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
prometheus_client==0.22.1
protobuf==5.29.5
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.16
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.33.0
sortedcontainers==2.4.0
soupsieve==2.7
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.46.1
streamlit-aggrid==1.1.6
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.1
streamlit-faker==0.0.3
streamlit-image-coordinates==0.3.1
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
